add fatigue.py for register fatigue to text file
